refactor: adjusted controllers

+++ database/migrations/2024_04_28_164533_create_accounts_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->float('balance');
+            $table->decimal('balance', 10);
             $table->boolean('active')->default(1);
             $table->timestamps();
         });
